// packages/dashboard-frontend/src/pages/UserPreferences/EnvironmentVariables/types.tsx
/*
 * Copyright (c) 2018-2024 Red Hat, Inc.
 * This program and the accompanying materials are made
 * available under the terms of the Eclipse Public License 2.0
 * which is available at https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *   Red Hat, Inc. - initial API and implementation
 */

import { api } from '@eclipse-che/common';

export type EditEnvironmentVariableProps =
  | {
  isEdit: true;
  variable: api.EnvironmentVariable;
}
  | {
  isEdit: false;
  variable: undefined;
};